requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
pandas==2.2.0
numpy==1.26.3
